pyproject.toml
[project]
name = "ai-financial-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.8.5",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "yfinance>=1.2.0",
]
